scripts/dataPromotion/requirements.txt
minio
python-dotenv